+++ packages/dive_obslib/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Changed method `outputGetState` to `streamOutputGetState`.
+- Added methods `streamOutputGetServiceNames` and `streamOutputGetServiceServers` to DiveFFIObslib
+to get service properties for streaming services.
+- Added `modify` method to DiveObslibData.
+- Added Int8Extensions.
+
 ## [0.1.0]
 ### Changed
 
@@ -23,7 +31,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added shutdowm method to DiveFFIObslib.
 - Added comments to code.
 - Added audio monitoring improvements.
-
 
 ## [0.0.1]
 ### Added